feat: Add logic to add new member to the ping group
This commit adds logic to the MemberJoinHandlerP3 feature to add a new member to the ping group. It checks if the member is already in the group and adds them if they are not. This ensures that the new member receives important notifications.

Refactor appMention to use regex
Add appMention feature to handle responding to app mentions in a channel
Add deleteMessage feature to handle deleting messages in a thread
Add logic to sign the guest book with user's message
Send last channel message in threads
Remove context and divider
Add MemberJoinHandlerP5 feature to handle announcing the presence of new members joining the channel
Add MemberJoinHandlerP4 feature to handle greeting the host for new members joining the channel
Add MemberJoinHandlerP3 feature to handle signing the guest book for new members joining the channel
Add memberJoinHandlerP2 feature to handle signing the guest book for new members joining the channel