feat: Add support for pinging @channel and announcing messages
This commit modifies the appMention feature to handle pinging the @channel group and announcing messages. It adds new cases to the switch statement to check for commands starting with "ping @channel" and "announce" and sends the appropriate message to the specified group. This enhancement improves the functionality of the appMention feature by allowing users to send important notifications to the entire channel.

Refactor appMention to use regex
Add appMention feature to handle responding to app mentions in a channel
Add deleteMessage feature to handle deleting messages in a thread
Add logic to sign the guest book with user's message
Send last channel message in threads
Remove context and divider
Add MemberJoinHandlerP5 feature to handle announcing the presence of new members joining the channel
Add MemberJoinHandlerP4 feature to handle greeting the host for new members joining the channel
Add MemberJoinHandlerP3 feature to handle signing the guest book for new members joining the channel
Add memberJoinHandlerP2 feature to handle signing the guest book for new members joining the channel