feat: Include user's name in the guest book message
This commit updates the memberJoinHandlerP3 feature to include the name of the new member in the guest book message. It retrieves the user's name using the Slack API and adds it to the markdown message. This enhancement improves the personalization of the guest book message for each new member.

Refactor appMention to use regex
Add appMention feature to handle responding to app mentions in a channel
Add deleteMessage feature to handle deleting messages in a thread
Add logic to sign the guest book with user's message
Send last channel message in threads
Remove context and divider
Add MemberJoinHandlerP5 feature to handle announcing the presence of new members joining the channel
Add MemberJoinHandlerP4 feature to handle greeting the host for new members joining the channel
Add MemberJoinHandlerP3 feature to handle signing the guest book for new members joining the channel
Add memberJoinHandlerP2 feature to handle signing the guest book for new members joining the channel