Refactor query time calculation in askChatGPT and hackerNews
This commit refactors the calculation of query time in the `askChatGPT` and `hackerNews` functions. It updates the formula to convert nanoseconds to seconds by dividing the result by 1,000,000,000 instead of 1,000,000. This ensures accurate representation of the query time in seconds. The change improves the precision of the query time measurement and provides more meaningful information to users.