chore: Update docker-compose.yaml with new image name and environment variable
This commit updates the `docker-compose.yaml` file to use the new image name `kcoderhtml/abot:latest` for the `server` service. It also adds the `OPENAI_API_KEY` environment variable with the value `sk-abot-dddddddddddddddddddddddddddddddd`. These changes ensure that the correct image is used and the necessary environment variable is set for the `abot` service.